requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
dacite==1.9.2
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.23.2
dvc==3.61.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvclive==3.48.3
entrypoints==0.4
filelock==3.18.0
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.7.2
htmlmin==0.1.12
hydra-core==1.3.2
idna==3.10
ImageHash==4.3.1
iterative-telemetry==0.0.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.6.3
multimethod==1.12
networkx==3.5
numba==0.61.0
numpy==2.1.3
nvidia-ml-py==12.575.51
omegaconf==2.3.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
patsy==1.0.1
phik==0.12.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
puremagic==1.30
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.1
Pygments==2.19.2
pygtrie==2.5.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
pywin32==311
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.15.3
scmrepo==3.4.0
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
statsmodels==0.14.5
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
visions==0.8.1
voluptuous==0.15.2
wcwidth==0.2.13
wordcloud==1.9.4
xgboost==3.0.3
yarl==1.20.1
ydata-profiling==4.16.1
zc.lockfile==3.0.post1
pytest